backend/pyproject.toml
[tool.poetry]
name = "learninggithub-backend"
version = "0.1.0"
description = "GitHub 项目学习助手后端服务"
authors = ["LearningGitHub Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
structlog = "^23.2.0"
# GitHub API
pygithub = "^2.1.1"
aiohttp = "^3.9.0"
# AI Integration (for future use)
openai = "^1.3.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
